+++ HISTORY.md
@@ -20,6 +20,7 @@
 * store: apply block migration dictionary to hidden blocks in microworlds
 * store: removed scaffolding
 * blocks: removed scaffolding
+* threads: removed scaffolding
 
 ### 2022-02-28
 * blocks, objects, threads, store: made addition reporter variadic